# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grooming CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentFull } from "./appointmentFull";

export interface DashboardSummary {
  todayAppointments: number;
  weekAppointments: number;
  monthRevenue: number;
  activeClients: number;
  pendingAppointments: number;
  recentAppointments?: AppointmentFull[];
}