# lib/api-zod/src/generated/types/petInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grooming CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { PetInputSize } from "./petInputSize";

export interface PetInput {
  clientId: number;
  name: string;
  breed?: string;
  size: PetInputSize;
  notes?: string;
}