# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grooming CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { ServiceSize } from "./serviceSize";

export interface Service {
  id: number;
  tenantId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  size: ServiceSize;
  price: number;
  /** @nullable */
  durationMinutes?: number | null;
  createdAt: Date;
}