# lib/api-zod/src/generated/types/topClient.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grooming CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface TopClient {
  clientId: number;
  clientName: string;
  totalRevenue: number;
  totalAppointments: number;
  /** @nullable */
  lastAppointmentDate?: string | null;
}